Move everything to base
We can start adding in things as we build out.
